% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{change_font_size}
\alias{change_font_size}
\title{Change the editor font size}
\usage{
change_font_size(pt)
}
\arguments{
\item{pt}{The font size (number) in points to change to in the editor.}
}
\value{

}
\description{
Change the font of the editor to be specific size.
}
\examples{
\dontrun{
change_font_size(12)
}
}
