% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{change_font_size_addin}
\alias{change_font_size_addin}
\title{Change font size addin}
\usage{
change_font_size_addin()
}
\value{

}
\description{
An RStudio addin that opens a selector window to choose the font size for the editor.
}
\examples{
\donttest{
change_font_size_addin()
}
}
