% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{decrement_font_size}
\alias{decrement_font_size}
\title{Decrement current font size}
\usage{
decrement_font_size(change = 2)
}
\arguments{
\item{change}{Number of points (positive number) to decrease the current font size by.}
}
\value{

}
\description{
Decrement current font size
}
\examples{
\dontrun{
decrement_font_size()

# decrease current font size by 6
decrement_font_size(change = 6)
}
}
