% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{get_current_editor_theme}
\alias{get_current_editor_theme}
\title{Get the current rstudio editor theme}
\usage{
get_current_editor_theme()
}
\value{
Name of the current rstudio theme being used.
}
\description{
Get the current rstudio editor theme
}
\examples{
\dontrun{
get_current_editor_theme
}
}
