% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachmode.R
\name{set_teach_mode_settings}
\alias{set_teach_mode_settings}
\title{Set the options for rstudioTeachMode}
\usage{
set_teach_mode_settings(
  current_mode = "user",
  user_theme_name = get_current_editor_theme(),
  user_font_size = get_current_font_size(),
  teach_theme_name = "Chrome",
  teach_font_size = 24,
  quiet = TRUE
)
}
\arguments{
\item{current_mode}{Either "user" or "teach".}

\item{user_theme_name}{Name of the theme to be used in "user" mode.}

\item{user_font_size}{Font size to be used in "user" mode.}

\item{teach_theme_name}{Name of the theme to be used in "teach" mode.}

\item{teach_font_size}{Font size to be used in "teach" mode.}

\item{quiet}{When FALSE prints the code to be added to your Rprofile to make settings persist between sessions.}
}
\value{
List object of the rstudioTeachMode options
}
\description{
Set the options for rstudioTeachMode
}
\examples{
\dontrun{
# Use current settings as the "user" mode and use the defaults for "teach" mode.
set_teach_mode_settings()
}
}
