% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toggle.R
\name{toggle_teach_mode}
\alias{toggle_teach_mode}
\title{Toggle between user preferences and teaching preferences}
\usage{
toggle_teach_mode()
}
\value{

}
\description{
Toggle between user preferences and teaching preferences
}
\examples{
\donttest{
toggle_teach_mode()
}
}
