% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{decrement_font_size}
\alias{decrement_font_size}
\title{Decrement current font size}
\usage{
decrement_font_size()
}
\value{

}
\description{
Decrement current font size
}
\examples{
\dontrun{
decrement_font_size()
}
}
