% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{increment_font_size}
\alias{increment_font_size}
\title{Increment current font size}
\usage{
increment_font_size()
}
\value{

}
\description{
Increment current font size
}
\examples{
\dontrun{
increment_font_size()
}
}
